// apps/web/form-builder/form-types.ts
type Input = {
  name: string;
  fieldType: "Input";
} & React.InputHTMLAttributes<HTMLInputElement> &
  SharedFormProps;

type PasswordInput = {
  name: string;
  fieldType: "Password";
  type: "password";
} & React.InputHTMLAttributes<HTMLInputElement> &
  SharedFormProps;

type FileUpload = {
  name: string;
  fieldType: "FileUpload";
  type: "file";
  maxSize: number;
  maxFiles?: number;
} & React.InputHTMLAttributes<HTMLInputElement> &
  SharedFormProps;

type OTPInput = {
  name: string;
  fieldType: "OTP";
} & OTPInputProps &
  SharedFormProps;

type Textarea = {
  name: string;
  fieldType: "Textarea";
} & React.TextareaHTMLAttributes<HTMLTextAreaElement> &
  SharedFormProps;

type Checkbox = {
  fieldType: "Checkbox";
} & CheckboxProps &
  SharedFormProps;

type RadioGroup = {
  fieldType: "RadioGroup";
  options: Option[];
} & RadioGroupProps &
  SharedFormProps;
//------------------------------
type ToggleGroupBaseProps = {
  fieldType: "ToggleGroup";
  options: Option[];
};

type ToggleGroupSingle = ToggleGroupBaseProps &
  ToggleGroupSingleProps & {
    type: "single";
  };

type ToggleGroupMultiple = ToggleGroupBaseProps &
  ToggleGroupMultipleProps & {
    type: "multiple";
  };

type ToggleGroup = (ToggleGroupSingle | ToggleGroupMultiple) & SharedFormProps;
//------------------------------

type Switch = {
  fieldType: "Switch";
} & SwitchProps &
  SharedFormProps;

type Slider = {
  fieldType: "Slider";
} & SliderProps &
  SharedFormProps;

type Select = {
  fieldType: "Select";
  /**
   * Options for the select field
   */
  options: Option[];
  placeholder: string;
} & React.SelectHTMLAttributes<HTMLSelectElement> &
  SharedFormProps;

type Combobox = {
  fieldType: "Combobox";
  /**
   * Options for the select field
   */
  options: Option[];
  placeholder: string;
} & React.SelectHTMLAttributes<HTMLSelectElement> &
  SharedFormProps;

type MultiSelect = {
  fieldType: "MultiSelect";
  /**
   * Options for the multiselect field
   */
  options: Option[];
  placeholder: string;
} & React.InputHTMLAttributes<HTMLInputElement> &
  SharedFormProps;
type DatePicker = {
  fieldType: "DatePicker";
  mode: "single" | "multiple" | "range";
} & React.InputHTMLAttributes<HTMLInputElement> &
  SharedFormProps;

type Rating = {
  fieldType: "Rating";
  numberOfStars?: number;
} & React.InputHTMLAttributes<HTMLInputElement> &
  SharedFormProps;

type TagInput = {
  fieldType: "TagInput";
  tags: { id: string; text: string }[];
} & TagInputProps &
  SharedFormProps;

/**
 * @deprecated
 */
type H1 = {
  fieldType: "H1";
  /**
   * the name is used as a key to identify the field
   */
  name: string;
  content: string;
  static: true;
} & React.HTMLAttributes<HTMLHeadingElement>;

/**
 * @deprecated
 */
type H2 = {
  fieldType: "H2";
  /**
   * the name is used as a key to identify the field
   */
  name: string;
  static: true;
  content: string;
} & React.HTMLAttributes<HTMLHeadingElement>;

/**
 * @deprecated
 */
type H3 = {
  fieldType: "H3";
  /**
   * the name is used as a key to identify the field
   */
  name: string;
  static: true;
  content: string;
} & React.HTMLAttributes<HTMLHeadingElement>;

/**
 * @deprecated
 */
type Paragraph = {
  fieldType: "P";
  /**
   * the name is used as a key to identify the field
   */
  name: string;
  static: true;
  content: string;
} & React.HTMLAttributes<HTMLParagraphElement>;

type Text = {
  fieldType: "Text";
  /**
   * the name is used as a key to identify the field
   */
  name: string;
  variant: "H1" | "H2" | "H3" | "P";
  static: true;
  content: string;
  width?: string;
} & React.HTMLAttributes<HTMLParagraphElement>;

type Divider = {
  fieldType: "Separator";
  /**
   * the name is used as a key to identify the field
   */
  name: string;
  label?: string;
  static: true;
} & SeparatorProps;

type SocialMediaButtons = {
  fieldType: "SocialMediaButtons";
  id: string;
  name: string;
  label: string;
  links: (keyof typeof socialLogsUrls)[];
  layout: "column" | "row";
};

/**
 * FormFieldType is a union type that represents all the possible form fields
 * that can be rendered in a form
 */
type FormFieldElement =
  | Textarea
  | Input
  | PasswordInput
  | OTPInput
  | Checkbox
  | RadioGroup
  | ToggleGroup
  | Switch
  | Select
  | MultiSelect
  | Slider
  | FileUpload
  | Rating
  | TagInput
  | Combobox
  | Text
  | SocialMediaButtons
  | DatePicker;

/**
 * StaticFormElement is a type that represents a static form element
 * that is not editable by the user
 */
export type StaticFormElement = H1 | H2 | H3 | Paragraph | Divider | Text;

export type FormElement =
  | (FormFieldElement & { id: string })
  | (StaticFormElement & { id: string });